bit68Task/requirements.txt
python3-openid==3.1.0
Django==2.1.5
django-allauth==0.39.1
django-crispy-forms==1.7.2
django-rest-auth==0.9.5
djangorestframework==3.9.2
requests==2.18.4


